refactor(native-rd): lift AnimatedSheet to its own component directory (#573)
Moves the shared bottom-sheet shell out of EvidenceTypePicker/, where five
call sites already deep-imported it from a directory it has nothing to do
with. Splits the scrim-free chrome out as SheetSurface so theme-matrix
stories can tile a body without stacking seven live scrims, and moves the
chrome style keys into AnimatedSheet.styles.ts so the shared shell no longer
imports styles from one specific consumer.

Pure move — no behaviour change, no test edits beyond the file location.
